Implement multi-line paste, conditionally on a Pike patch.
As this patch has yet to be accepted into trunk, it may be modified before going
live. If so, this code will be changed to match it, but that'll simply mean that
this version of the Gypsum end of the code will never trip (eg if the function's
called something different).