Big one: Merge plugins-more into plugins.
Configuring of plugins, discovery of active-by-default, etc is now in window
rather than in plugins/more (which no longer exists). All the files that were
in -more are now in the main plugins directory, and any plugin can be unloaded
via the configuration menu. You can't shoot yourself in the foot that way, as
you can always reload a plugin that same way.