window: Shift the Pause handling out of the EF keypress into a menu item.
Improves visibility (it's there as a menu item, not just a magic key), and
makes it work when focus isn't on the primary EF (eg when you're in Ctrl-F
search). Downside of the visibility improvement is that the Windows issue,
where VoidSymbol is used instead of Pause, is now equally visible.