RMImport: Revert the last couple of commits and correct a comment.
Whoops... the RosMud feature "inputcol" is nothing to do with the Gypsum
feature of the same name. RosMud allows the input field to be either colored
(using the specified foreground and background) or uncolored (using whatever
styles Windows provides by default). Gypsum's feature is quite different: it's
the color used in scrollback for inputted text. So importing this makes no
sense, and shouldn't be done.