Fix docs bug - the x,y variables haven't been used properly in practically forever.
This is one of those big arguables. Am I removing a pledge? The functionality
was present for exactly one commit, and was defanged immediately afterward;
movablewindow was implemented in 0b6266, and the retention of size was added
one commit later, in e80223, and the size retention code broke the defaulting
via x,y. It was documented, however, and according to the 1.0.0 tag, ought to
be a part of every Gypsum less than 2.0; but in order to comply with that, I
would need to first implement the behaviour, and then immediately deprecate it
- not the best plan for code maintenance. So I'm just removing this. It's not
possible for anyone to have been depending on it, as it fundamentally did not
work; so I'm calling this a bugfix. But it's definitely arguable.