window: And finally, after all the preliminaries... merge movablewindow and window.pike!!
As will have been obvious to anyone who's been reading the commits of the last
24 hours, this is done by having window.pike inherit movablewindow; this is by
far the cleanest way to do things. The only part of this that isn't really the
clean code that I normally like to do is in window.pike::create, where parents
are split out - instead of just calling ::create(), it first calls the crucial
movablewindow::create(), and then when that's fully set up, it calls the other
::create functions (at the moment, that's just one - for the status bar slot -
but that may grow in future, and the code's not going to be affected by adding
a third inherit). So, this is the Good Thing that I've been working towards :)