Bugfix statustext_maxwidth at the expense of performance.
Instead of asking the current label how wide it'd be, which could be affected
by a previous set_size_request(), create a new GTK2.Label and ask it about the
necessary width. This is less efficient, but fixes the issue with the label
sometimes not expanding as it ought to.