connection: Use makeconn for passive connections.
It's the moment you've all been waiting for, folks! :) This unifies two pieces
of code, and notably, fixes a bug in sockaccept() wherein it would emit a null
color object, support for which has since been dropped. It also allows further
improvements in the future... some of which will be the extremely near future.