update: Work around an oddity.
For some reason, just now, there's something weird happening: using -= here,
which ought to be absolutely identical to the code I'm now using, utterly
fails. It just leaves globalusage as it is. It's as if something's getting
optimized out, perhaps. It's hard to craft a test-case, though; doing the
exact same thing from the /x command line works just fine.