update: Update comment now that I know more about the cause of this issue.
It's an optimization gone awry, so hopefully one of the Pike core devs can
figure out what's going on. In the meantime, this works fine; and even if a
faulty Pike is used, all it means is that unloading is imperfect. Not a big
deal, ergo not worth guarding with COMPAT checks.