Experimentally shift the Save and Delete buttons to inside the configdlg.
Lots of whitespace changes, not a lot of structural change, but it makes one
significant shift to the window layout of configdlgs. Instead of having Save,
Delete, and Close, all as equals, there's now Save and Delete at the bottom of
the make_content() blob (note that this will at times be a bit redundant, as a
configdlg will often create a Vbox for its content - but in case it doesn't, a
dedicated Vbox is used for the outside), with Close on its own down below. An
additional button goes there too (eg Options|Keyboard), for better or worse.