Alias: Make world aliases trump global aliases. BREAKING CHANGE.
I doubt that anyone will have conflicting aliases, so this probably won't
actually break anything for anyone. This is the more logical way to prioritize
aliases; the more specific trumps the global, so you can have a per-world
alias for each of several worlds, and a global one for all other worlds. The
previous behaviour means that per-world aliases are utterly ignored if there's
a global alias on the same word.