window: Make use of collect_signals(). Note that this is an API change.
Previously, subwsignals() could be called at any time for any subw, and it
would replace all that subw's signals. Now, it consists of two parts: one that
functions as it previously did, and one that appends to win->signals, so it
has to be done only for a new subw or after signals have just been wiped. As
these are the only two times subwsignals() is called anyway, this is not a
problem, but it still creates a strange transitional state.