window: Rename charwidth to enwidth.
Originally, charwidth was the expected average width of any character (well,
before that it was the expected _actual_ width, but that was doomed from the
start), but its purpose has been eroded until it's just for setting tab width
(a tab is defined as eight ens). So now we define it that way.