window: Implement block selection as described
Turns out it's not as hard as I thought. Seeing what Open Office Writer did
gave me the idea for implementation, and voila, there it is. There may still
be performance issues, but that just means "don't use block selection on low
end systems", and the feature should be fine.